% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex1.R
\docType{data}
\name{ex1}
\alias{ex1}
\title{Dados de exemplo de um experimento em DQL.}
\format{
Simulated example dataset for analysis of split-split-plot design
(Conjunto de dados simulados de exemplo para a analise do esquema de parcelas subsubdivididas).
}
\usage{
data(ex1)
}
\description{
Exemplo simulado com um conjunto de dados
de um experimento em quadrados latino
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK72M3lFl7kBaHiVh5W53mlR>
}
\author{
Alcinei Mistico Azevedo: \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
