% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitSplitPlot-package.R
\docType{package}
\name{SplitSplitPlot-package}
\alias{SplitSplitPlot-package}
\alias{SplitSplitPlot}
\title{Analysis of split-split-plot experiments (Analise de experimentos em parcela subsubdividida)}
\description{
SplitSplitPlot: Performs analysis of split-split-plot experiments(realiza a analise de experimentos em
parcela subsubdividida).
}
\references{
Canal no youtube -> Alcinei Azevedo - Dicas e aulas
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>

 BANZATTO, D. A.; KRONKA, S. N. Experimentacao Agricola. 4 ed.
 Jaboticabal: Funep. 2006. 237 p.

 GOMES, F. P. Curso de Estatistica Experimental. 10a ed. Piracicaba:
 ESALQ/USP. 1982. 430.
}
\seealso{
\code{\link{aov}}, \code{\link{lm}}
}
\author{
Alcinei Mistico Azevedo <alcineimistico@hotmail.com>
}
